+++ BoosterPumpLibrary/Settings/Register.cs
@@ -2,7 +2,7 @@
 {
     public class Register : RegisterBase<ulong>
     {
-        protected override int MaxSize => 8;
+        protected override ushort MaxSize => 8;
 
         /// <summary>
         /// Describes a register in a I2C device with one or multiple bytes (max 8).
@@ -10,7 +10,7 @@
         /// <param name="registerAddress"></param>
         /// <param name="description"></param>
         /// <param name="byteCount"></param>
-        public Register(byte registerAddress, string description, int byteCount) : base(registerAddress, description, byteCount)
+        public Register(byte registerAddress, string description, ushort byteCount) : base(registerAddress, description, byteCount)
         { }              
              
         internal override void SetValue(ulong value)
